chore: clean up server config
